requirements.txt
langchain-core
langchain
langchain-chroma
langchain-community
langchain-openai
langchain-text-splitters
Flask==3.0.3
tensorflow==2.16.1
numpy==1.26.4
python-dotenv
-e .






